postprocessing moved to single script what runs on entire segmentation directory sequentially
